Move qgis_messaging.py into views/components
