+++ arches_project/core/arches/connection.py
@@ -2,7 +2,7 @@ import requests
 from datetime import datetime, timedelta
 
 from arches_project.core.views.login import LoggedIn
-from arches_project.core.utils.qgis_messaging import show_message
+from arches_project.core.views.components.qgis_messaging import show_message
 from arches_project.core.views.components.spinner import triggerSpinner
 
 from arches_project.core.arches.api import arches_api
